# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  '/api/v1/images/profile/{profile_id}':
    get:
      summary: ''
      operationId: getApiV1ImagesProfileProfile_id
      description: ''
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: true
                  message: 'profile Images for user:Dr. Orlo Bradtke'
                  data:
                    images: []
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'profile Images for user:Dr. Orlo Bradtke'
                  data:
                    type: object
                    example:
                      images: []
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: profile_id
        description: 'The ID of the profile.'
        example: 1
        required: true
        schema:
          type: integer
...
  '/api/v1/files/{postFile_id}':
    get:
      summary: ''
      operationId: getApiV1FilesPostFile_id
      description: ''
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: true
                  message: file
                  data:
                    file:
                      id: 1
                      post_id: 57
                      path: 'https://source.unsplash.com/random'
                      type: image
                      created_at: '2022-11-16T19:25:31.000000Z'
                      updated_at: '2022-11-16T19:25:31.000000Z'
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: file
                  data:
                    type: object
                    example:
                      file:
                        id: 1
                        post_id: 57
                        path: 'https://source.unsplash.com/random'
                        type: image
                        created_at: '2022-11-16T19:25:31.000000Z'
                        updated_at: '2022-11-16T19:25:31.000000Z'
      tags:
        - Endpoints
      security: []
    delete:
      summary: ''
      operationId: deleteApiV1FilesPostFile_id
      description: ''
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: postFile_id
        description: 'The ID of the postFile.'
        example: 1
        required: true
        schema:
          type: integer
...
  '/api/v1/users/{user_id}':
    get:
      summary: 'show user'
      operationId: showUser
      description: ''
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: true
                  message: User
                  data:
                    user:
                      id: 1
                      first_name: Otto
                      last_name: Mitchell
                      user_name: 'Dr. Orlo Bradtke'
                      phone: '+1 (678) 832-8771'
                      age: 25
                      gender: female
                      email: adicki@example.org
                      email_verified_at: '2022-11-16T19:25:27.000000Z'
                      created_at: '2022-11-16T19:25:30.000000Z'
                      updated_at: '2022-11-16T19:25:30.000000Z'
                      following:
                        -
                          id: 15
                          user_id: 15
                          description: 'Quo saepe velit blanditiis ipsa sed optio atque.'
                          avatar: 'https://source.unsplash.com/random'
                          cover: 'https://source.unsplash.com/random'
                          created_at: '2022-11-16T19:25:30.000000Z'
                          updated_at: '2022-11-16T19:25:30.000000Z'
                          pivot:
                            user_id: 1
                            profile_id: 15
                        -
                          id: 18
                          user_id: 18
                          description: 'Alias et odit iste aliquid quos et ut.'
                          avatar: 'https://source.unsplash.com/random'
                          cover: 'https://source.unsplash.com/random'
                          created_at: '2022-11-16T19:25:30.000000Z'
                          updated_at: '2022-11-16T19:25:31.000000Z'
                          pivot:
                            user_id: 1
                            profile_id: 18
                        -
                          id: 20
                          user_id: 20
                          description: 'Tempora est qui optio optio quo.'
                          avatar: 'https://source.unsplash.com/random'
                          cover: 'https://source.unsplash.com/random'
                          created_at: '2022-11-16T19:25:30.000000Z'
                          updated_at: '2022-11-16T19:25:31.000000Z'
                          pivot:
                            user_id: 1
                            profile_id: 20
                        -
                          id: 39
                          user_id: 39
                          description: 'Esse magni aut expedita consequatur enim nesciunt distinctio.'
                          avatar: 'https://source.unsplash.com/random'
                          cover: 'https://source.unsplash.com/random'
                          created_at: '2022-11-16T19:25:30.000000Z'
                          updated_at: '2022-11-16T19:25:31.000000Z'
                          pivot:
                            user_id: 1
                            profile_id: 39
                        -
                          id: 48
                          user_id: 48
                          description: 'Laboriosam ea provident eum quo vel consequatur voluptatum.'
                          avatar: 'https://source.unsplash.com/random'
                          cover: 'https://source.unsplash.com/random'
                          created_at: '2022-11-16T19:25:30.000000Z'
                          updated_at: '2022-11-16T19:25:31.000000Z'
                          pivot:
                            user_id: 1
                            profile_id: 48
                      profile:
                        id: 1
                        user_id: 1
                        description: 'Eius quis harum in voluptates.'
                        avatar: 'https://source.unsplash.com/random'
                        cover: 'https://source.unsplash.com/random'
                        created_at: '2022-11-16T19:25:30.000000Z'
                        updated_at: '2022-11-16T19:25:30.000000Z'
                        followers:
                          -
                            id: 19
                            first_name: Pascale
                            last_name: Murazik
                            user_name: 'Cullen Schumm'
                            phone: +1-651-553-6779
                            age: 39
                            gender: female
                            email: layne.prohaska@example.net
                            email_verified_at: '2022-11-16T19:25:28.000000Z'
                            created_at: '2022-11-16T19:25:30.000000Z'
                            updated_at: '2022-11-16T19:25:30.000000Z'
                            pivot:
                              profile_id: 1
                              user_id: 19
                          -
                            id: 24
                            first_name: Saul
                            last_name: Howe
                            user_name: 'Name Jerde DVM'
                            phone: 757-476-5347
                            age: 53
                            gender: female
                            email: torphy.wilford@example.org
                            email_verified_at: '2022-11-16T19:25:28.000000Z'
                            created_at: '2022-11-16T19:25:30.000000Z'
                            updated_at: '2022-11-16T19:25:30.000000Z'
                            pivot:
                              profile_id: 1
                              user_id: 24
                          -
                            id: 42
                            first_name: Stephen
                            last_name: Hartmann
                            user_name: 'Eveline Bogan'
                            phone: 856-304-0563
                            age: 56
                            gender: male
                            email: edmond.kilback@example.org
                            email_verified_at: '2022-11-16T19:25:29.000000Z'
                            created_at: '2022-11-16T19:25:30.000000Z'
                            updated_at: '2022-11-16T19:25:30.000000Z'
                            pivot:
                              profile_id: 1
                              user_id: 42
                          -
                            id: 16
                            first_name: Nathanial
                            last_name: Witting
                            user_name: 'Zita Collier'
                            phone: '+18786460812'
                            age: 26
                            gender: female
                            email: giovanny16@example.net
                            email_verified_at: '2022-11-16T19:25:28.000000Z'
                            created_at: '2022-11-16T19:25:30.000000Z'
                            updated_at: '2022-11-16T19:25:30.000000Z'
                            pivot:
                              profile_id: 1
                              user_id: 16
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: User
                  data:
                    type: object
                    example:
                      user:
                        id: 1
                        first_name: Otto
                        last_name: Mitchell
                        user_name: 'Dr. Orlo Bradtke'
                        phone: '+1 (678) 832-8771'
                        age: 25
                        gender: female
                        email: adicki@example.org
                        email_verified_at: '2022-11-16T19:25:27.000000Z'
                        created_at: '2022-11-16T19:25:30.000000Z'
                        updated_at: '2022-11-16T19:25:30.000000Z'
                        following:
                          -
                            id: 15
                            user_id: 15
                            description: 'Quo saepe velit blanditiis ipsa sed optio atque.'
                            avatar: 'https://source.unsplash.com/random'
                            cover: 'https://source.unsplash.com/random'
                            created_at: '2022-11-16T19:25:30.000000Z'
                            updated_at: '2022-11-16T19:25:30.000000Z'
                            pivot:
                              user_id: 1
                              profile_id: 15
                          -
                            id: 18
                            user_id: 18
                            description: 'Alias et odit iste aliquid quos et ut.'
                            avatar: 'https://source.unsplash.com/random'
                            cover: 'https://source.unsplash.com/random'
                            created_at: '2022-11-16T19:25:30.000000Z'
                            updated_at: '2022-11-16T19:25:31.000000Z'
                            pivot:
                              user_id: 1
                              profile_id: 18
                          -
                            id: 20
                            user_id: 20
                            description: 'Tempora est qui optio optio quo.'
                            avatar: 'https://source.unsplash.com/random'
                            cover: 'https://source.unsplash.com/random'
                            created_at: '2022-11-16T19:25:30.000000Z'
                            updated_at: '2022-11-16T19:25:31.000000Z'
                            pivot:
                              user_id: 1
                              profile_id: 20
                          -
                            id: 39
                            user_id: 39
                            description: 'Esse magni aut expedita consequatur enim nesciunt distinctio.'
                            avatar: 'https://source.unsplash.com/random'
                            cover: 'https://source.unsplash.com/random'
                            created_at: '2022-11-16T19:25:30.000000Z'
                            updated_at: '2022-11-16T19:25:31.000000Z'
                            pivot:
                              user_id: 1
                              profile_id: 39
                          -
                            id: 48
                            user_id: 48
                            description: 'Laboriosam ea provident eum quo vel consequatur voluptatum.'
                            avatar: 'https://source.unsplash.com/random'
                            cover: 'https://source.unsplash.com/random'
                            created_at: '2022-11-16T19:25:30.000000Z'
                            updated_at: '2022-11-16T19:25:31.000000Z'
                            pivot:
                              user_id: 1
                              profile_id: 48
                        profile:
                          id: 1
                          user_id: 1
                          description: 'Eius quis harum in voluptates.'
                          avatar: 'https://source.unsplash.com/random'
                          cover: 'https://source.unsplash.com/random'
                          created_at: '2022-11-16T19:25:30.000000Z'
                          updated_at: '2022-11-16T19:25:30.000000Z'
                          followers:
                            -
                              id: 19
                              first_name: Pascale
                              last_name: Murazik
                              user_name: 'Cullen Schumm'
                              phone: +1-651-553-6779
                              age: 39
                              gender: female
                              email: layne.prohaska@example.net
                              email_verified_at: '2022-11-16T19:25:28.000000Z'
                              created_at: '2022-11-16T19:25:30.000000Z'
                              updated_at: '2022-11-16T19:25:30.000000Z'
                              pivot:
                                profile_id: 1
                                user_id: 19
                            -
                              id: 24
                              first_name: Saul
                              last_name: Howe
                              user_name: 'Name Jerde DVM'
                              phone: 757-476-5347
                              age: 53
                              gender: female
                              email: torphy.wilford@example.org
                              email_verified_at: '2022-11-16T19:25:28.000000Z'
                              created_at: '2022-11-16T19:25:30.000000Z'
                              updated_at: '2022-11-16T19:25:30.000000Z'
                              pivot:
                                profile_id: 1
                                user_id: 24
                            -
                              id: 42
                              first_name: Stephen
                              last_name: Hartmann
                              user_name: 'Eveline Bogan'
                              phone: 856-304-0563
                              age: 56
                              gender: male
                              email: edmond.kilback@example.org
                              email_verified_at: '2022-11-16T19:25:29.000000Z'
                              created_at: '2022-11-16T19:25:30.000000Z'
                              updated_at: '2022-11-16T19:25:30.000000Z'
                              pivot:
                                profile_id: 1
                                user_id: 42
                            -
                              id: 16
                              first_name: Nathanial
                              last_name: Witting
                              user_name: 'Zita Collier'
                              phone: '+18786460812'
                              age: 26
                              gender: female
                              email: giovanny16@example.net
                              email_verified_at: '2022-11-16T19:25:28.000000Z'
                              created_at: '2022-11-16T19:25:30.000000Z'
                              updated_at: '2022-11-16T19:25:30.000000Z'
                              pivot:
                                profile_id: 1
                                user_id: 16
      tags:
        - User
      security: []
    parameters:
      -
        in: path
        name: user_id
        description: 'The ID of the user.'
        example: 1
        required: true
        schema:
          type: integer
  '/api/v1/users/{user_id}/profile':
    get:
      summary: 'show profile for user'
      operationId: showProfileForUser
      description: ''
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: true
                  message: 'Profile for User'
                  data:
                    profile:
                      id: 1
                      user_id: 1
                      description: 'Eius quis harum in voluptates.'
                      avatar: 'https://source.unsplash.com/random'
                      cover: 'https://source.unsplash.com/random'
                      created_at: '2022-11-16T19:25:30.000000Z'
                      updated_at: '2022-11-16T19:25:30.000000Z'
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Profile for User'
                  data:
                    type: object
                    example:
                      profile:
                        id: 1
                        user_id: 1
                        description: 'Eius quis harum in voluptates.'
                        avatar: 'https://source.unsplash.com/random'
                        cover: 'https://source.unsplash.com/random'
                        created_at: '2022-11-16T19:25:30.000000Z'
                        updated_at: '2022-11-16T19:25:30.000000Z'
      tags:
        - User
      security: []
    parameters:
      -
        in: path
        name: user_id
        description: 'The ID of the user.'
        example: 1
        required: true
        schema:
          type: integer
...
components:
  securitySchemes:
    default:
      type: apiKey
      name: apiKey
      in: query
      description: ''
security:
  -
    default: []